I did a lot
After playtesting last night, I re-added the fill in gaps method from the original SongClash, and that seemed to work well. I also added a mostly-functional timer, and the ability to be re-kicked back to start. I also changed the shuffle method to not include songs without preview audio, to enhance the game experience. Feel pretty good about this one.